+++ lib/constants/app_glob_setting.dart
@@ -3,9 +3,9 @@ import 'package:env/env.dart';
 class AppGlobalSetting {
   const AppGlobalSetting._();
 
-  static final baseUrl = EnvFlavor().baseUrl;
+  static final baseUrl = EnvFlavor().baseUrl ?? '';
 
-  static final tmdbUrl = EnvFlavor().tmdbUrl;
+  static final tmdbUrl = EnvFlavor().tmdbUrl ?? '';
 
-  static final tmdbImg = EnvFlavor().tmdbImg;
+  static final tmdbImg = EnvFlavor().tmdbImg ?? '';
 }